[useTree]: refactored and fixed form save.
