+++ uui-core/src/data/processing/views/tree/Tree.ts
@@ -1,9 +1,8 @@
 import isEqual from 'lodash.isequal';
-import { CascadeSelection, CascadeSelectionTypes, DataRowPathItem, DataSourceState, IMap, LazyDataSourceApi } from '../../../../../types';
+import { CascadeSelection, CascadeSelectionTypes, DataRowPathItem, DataSourceState, IMap, LazyDataSourceApi } from '../../../../types';
 import { ITree } from './ITree';
-import { FULLY_LOADED, NOT_FOUND_RECORD } from '../constants';
-import { FetchingHelper } from './treeStructure';
-import { ITreeNodeInfo } from './exposed';
+import { FULLY_LOADED, NOT_FOUND_RECORD } from './constants';
+import { FetchingHelper, ITreeNodeInfo } from './treeStructure';
 
 export interface LoadOptions<TItem, TId, TFilter = any> {
     tree: ITree<TItem, TId>;
