+++ uui-core/src/data/processing/views/tree/exposed.ts
@@ -1,0 +1,17 @@
+export * from './treeStructure/exposed';
+export type { ITree, ITreeItemsInfo, ITreeNodeStatus } from './ITree';
+export { Tree } from './Tree';
+export type { ITreeLoadResult } from './Tree';
+export {
+    NOT_FOUND_RECORD,
+
+    PARTIALLY_LOADED,
+    FULLY_LOADED,
+    EMPTY,
+
+    FAILED_RECORD,
+    PENDING_RECORD,
+    LOADING_RECORD,
+    LOADED_RECORD,
+} from './constants';
+export { CompositeKeysMap, newMap, cloneMap } from './helpers';
