+++ uui-core/src/data/processing/views/tree/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './comparators';
+export * from './map';
