+++ uui-core/src/data/processing/views/tree/hooks/common/useSearchTree.ts
@@ -1,6 +1,6 @@
 import { useSimplePrevious } from '../../../../../../hooks';
 import { DataSourceState } from '../../../../../../types';
-import { TreeState } from '../../newTree';
+import { TreeState } from '../../treeState';
 import { useUpdateTree } from './useUpdateTree';
 
 export type UseSearchTreeProps<TItem, TId, TFilter = any> = {
