+++ uui-core/src/data/processing/views/tree/hooks/common/useUpdateTree.ts
@@ -1,5 +1,5 @@
 import { useMemo, useRef } from 'react';
-import { TreeState } from '../../newTree';
+import { TreeState } from '../../treeState';
 import { useSimplePrevious } from '../../../../../../hooks';
 import { useDepsChanged } from './useDepsChanged';
