+++ uui-core/src/data/processing/views/tree/hooks/strategies/lazyTree/__tests__/useLazyTree.test.ts
@@ -3,9 +3,10 @@ import { useLazyTree } from '../useLazyTree';
 import { DataQueryFilter, DataSourceState } from '../../../../../../../../types';
 import { LocationItem } from '../../../../../../__tests__/mocks';
 import { demoData } from '@epam/uui-docs';
-import { FAILED_RECORD, NOT_FOUND_RECORD, TreeStructure, newMap } from '../../../../newTree';
+import { FAILED_RECORD, NOT_FOUND_RECORD, newMap } from '../../../../exposed';
 import { ItemsStorage } from '../../../../ItemsStorage';
 import { RecordStatus } from '../../../../types';
+import { TreeStructure } from '../../../../treeStructure';
 
 describe('useLazyTree', () => {
     let dataSourceState: DataSourceState<DataQueryFilter<LocationItem>, string>;
