+++ uui-core/src/data/processing/views/tree/index.ts
@@ -1,3 +1,3 @@
 export * from './hooks';
-export * from './newTree/exposed';
+export * from './exposed';
 export * from './PatchOrderingMap';
