--- uui-core/src/data/processing/views/tree/newTree/exposed.ts
@@ -1,5 +1,0 @@
-export * from './treeStructure/exposed';
-export type { ITree, ITreeItemsInfo, ITreeNodeStatus } from './ITree';
-export { Tree } from './Tree';
-export type { ITreeLoadResult } from './Tree';
-export { NOT_FOUND_RECORD } from '../constants';
