--- uui-core/src/data/processing/views/tree/newTree/index.ts
@@ -1,3 +1,0 @@
-export * from '../constants';
-export * from './treeState';
-export * from './treeStructure';
