--- uui-core/src/data/processing/views/tree/newTree/treeStructure/exposed.ts
@@ -1,2 +1,0 @@
-export { FetchingHelper, CompositeKeysMap } from './helpers';
-export type { ITreeNodeInfo, ITreeParams } from './types';
