+++ uui-core/src/data/processing/views/tree/treeStructure/exposed.ts
@@ -1,0 +1,2 @@
+export { FetchingHelper } from './helpers';
+export type { ITreeNodeInfo, ITreeParams } from './types';
