+++ uui-core/src/data/processing/views/tree/treeStructure/helpers/index.ts
@@ -3,6 +3,5 @@ export * from './FetchingHelper';
 export * from './FilterHelper';
 export * from './SearchHelper';
 export * from './SortHelper';
-export * from './map';
 export * from './checked';
 export * from './PatchHelper';
