+++ uui-core/src/data/processing/views/tree/treeStructure/index.ts
@@ -1,2 +1,3 @@
 export * from './TreeStructure';
 export * from './helpers';
+export * from './types';
